feat: move types to own folder
